refactor: Move primitives into files
